refactor into simulation.go
