--- x/participation/simulation/participate.go
@@ -1,32 +1,0 @@
-package simulation
-
-import (
-	"math/rand"
-
-	"github.com/cosmos/cosmos-sdk/baseapp"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
-	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
-
-	"github.com/tendermint/spn/x/participation/keeper"
-	"github.com/tendermint/spn/x/participation/types"
-)
-
-func SimulateMsgParticipate(
-	_ authkeeper.AccountKeeper,
-	_ bankkeeper.Keeper,
-	_ keeper.Keeper,
-) simtypes.Operation {
-	return func(r *rand.Rand, app *baseapp.BaseApp, ctx sdk.Context, accs []simtypes.Account, chainID string,
-	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
-		simAccount, _ := simtypes.RandomAcc(r, accs)
-		msg := &types.MsgParticipate{
-			Participant: simAccount.Address.String(),
-		}
-
-		// TODO: Handling the Participate simulation
-
-		return simtypes.NoOpMsg(types.ModuleName, msg.Type(), "Participate simulation not implemented"), nil, nil
-	}
-}
