+++ x/participation/simulation/simulation.go
@@ -13,6 +13,24 @@ import (
 	"github.com/tendermint/spn/x/participation/types"
 )
 
+func SimulateMsgParticipate(
+	_ authkeeper.AccountKeeper,
+	_ bankkeeper.Keeper,
+	_ keeper.Keeper,
+) simtypes.Operation {
+	return func(r *rand.Rand, app *baseapp.BaseApp, ctx sdk.Context, accs []simtypes.Account, chainID string,
+	) (simtypes.OperationMsg, []simtypes.FutureOperation, error) {
+		simAccount, _ := simtypes.RandomAcc(r, accs)
+		msg := &types.MsgParticipate{
+			Participant: simAccount.Address.String(),
+		}
+
+		// TODO: Handling the Participate simulation
+
+		return simtypes.NoOpMsg(types.ModuleName, msg.Type(), "Participate simulation not implemented"), nil, nil
+	}
+}
+
 func SimulateMsgWithdrawAllocations(
 	_ authkeeper.AccountKeeper,
 	_ bankkeeper.Keeper,
